+++ imager/src/exfat.rs
@@ -4,11 +4,12 @@ use std::mem;
 use std::path;
 
 mod boot_sector;
+mod extended_boot_sector;
 
 #[derive(Debug)]
 pub struct Exfat {
     boot_sector: boot_sector::BootSector,
-    extended_boot_sectors: [ExtendedBootSector; 0x8],
+    extended_boot_sectors: [extended_boot_sector::ExtendedBootSector; 0x8],
     oem_parameter_sector: OemParameterSector,
     reserved_sector: ReservedSector,
     boot_checksum_sector: Option<BootChecksumSector>,
@@ -19,7 +20,7 @@ impl Exfat {
         let boot_sector = boot_sector::BootSector::new(&boot_sector);
         Self {
             boot_sector,
-            extended_boot_sectors: [ExtendedBootSector::new(); 0x8],
+            extended_boot_sectors: [extended_boot_sector::ExtendedBootSector::new(); 0x8],
             oem_parameter_sector: OemParameterSector::null_parameters(),
             reserved_sector: ReservedSector::new(),
             boot_checksum_sector: None,
@@ -100,77 +101,6 @@ trait Packable {
 trait Unpackable {
     type Unpacked;
     fn unpack(&self) -> Self::Unpacked;
-}
-
-#[derive(Clone, Copy, Debug)]
-struct ExtendedBootSector {
-    boot_code: [u8; 0x1fc],
-    boot_signature: u32,
-}
-
-impl ExtendedBootSector {
-    fn new() -> Self {
-        Self {
-            boot_code: [0; 0x1fc],
-            boot_signature: 0xaa550000,
-        }
-    }
-}
-
-impl Packable for ExtendedBootSector {
-    type Packed = PackedExtendedBootSector;
-
-    fn pack(&self) -> Self::Packed {
-        Self::Packed {
-            boot_code: self.boot_code,
-            boot_signature: self.boot_signature,
-        }
-    }
-}
-
-impl Sector for ExtendedBootSector {
-    fn to_bytes(&self) -> RawSector {
-        self.pack().to_bytes()
-    }
-}
-
-impl fmt::Display for ExtendedBootSector {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "extended_boot_sector.boot_code = {:x?}\n", self.boot_code)?;
-        write!(f, "extended_boot_sector.boot_signature = {:x?}", self.boot_signature)
-    }
-}
-
-#[derive(Clone, Copy)]
-#[repr(packed)]
-struct PackedExtendedBootSector {
-    boot_code: [u8; 0x1fc],
-    boot_signature: u32,
-}
-
-impl Unpackable for PackedExtendedBootSector {
-    type Unpacked = ExtendedBootSector;
-
-    fn unpack(&self) -> Self::Unpacked {
-        Self::Unpacked {
-            boot_code: self.boot_code,
-            boot_signature: self.boot_signature,
-        }
-    }
-}
-
-impl Sector for PackedExtendedBootSector {
-    fn to_bytes(&self) -> RawSector {
-        unsafe {
-            mem::transmute::<Self, [u8; mem::size_of::<Self>()]>(*self)
-        }
-    }
-}
-
-impl fmt::Display for PackedExtendedBootSector {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        self.unpack().fmt(f)
-    }
 }
 
 #[derive(Clone, Copy, Debug)]
